% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagConnect.R
\name{api_auth_header}
\alias{api_auth_header}
\title{An S3 class representing the API authentication for a Tag.bio server}
\usage{
api_auth_header(object, ...)
}
\arguments{
\item{object}{tagConnect object}

\item{...}{future extensions}
}
\description{
The api_auth_header class establishes models the authentication token
for a local or remote Tag.bio server.
}
\section{API Authentication Notes}{

There are three different modes for authenticating to the API:
\itemize{
\item 1. For an API call connected via localhost, no authentication is
   required.  Localhost is the assumed endpoint if:
   - no URL is provided
   - URL contains "localhost" or "127.0.0.1"
\item 2. When an API key is provided, connection is made using basic
   authentication with user name and password extracted from the
   API key.
\item 3. If neither, but a token is provided, use token
   authentication.
}
}

