% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagConnect.R
\name{api_auth_header.tagConnect}
\alias{api_auth_header.tagConnect}
\title{An S3 class representing a connection to a Tag.bio server}
\usage{
\method{api_auth_header}{tagConnect}(object, url)
}
\arguments{
\item{url}{URL to connect to}

\item{api_auth_header}{object}
}
\description{
The tagConnect class establishes a connection to a local or remote
Tag.bio server.
}
