% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagConnect.R
\name{api_get}
\alias{api_get}
\title{An S3 class representing a Tag.bio API data product getter}
\usage{
api_get(object, ...)
}
\arguments{
\item{tagConnect}{object}
}
\description{
The tagConnect class establishes a connection to a local or remote
Tag.bio server.
}
\details{
Connects to a data product via a tagConnect connection.
}
\examples{
# Connected to Tag.bio data host using API key

}
