% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagConnect.R
\name{api_get.tagConnect}
\alias{api_get.tagConnect}
\title{Tag.bio API data product getter}
\usage{
\method{api_get}{tagConnect}(object, url)
}
\arguments{
\item{api_auth_header}{object}
}
\description{
Connects to a data product via a tagConnect connection.
}
\examples{
# Connected to Tag.bio data host using API key

}
