% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagConnect.R
\name{api_post.tagConnect}
\alias{api_post.tagConnect}
\title{Tag.bio API data product post}
\usage{
\method{api_post}{tagConnect}(object, query_type, url, return_type = "json", jsonPayload = NA)
}
\arguments{
\item{query_type}{type of query}

\item{url}{URL to connect to}

\item{return_type}{return type ("json" or "table")}

\item{jsonPayload}{JSON payload for POST}

\item{api_auth_header}{object}
}
\description{
Connects to a data product via a tagConnect connection.
}
\examples{
\dontrun{
# Connected to Tag.bio data host using API key
}

}
