% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_fundamentals.R
\name{CategoricalBatch}
\alias{CategoricalBatch}
\title{An S3 class representing the Tag.bio CategoricalBatch}
\usage{
CategoricalBatch(collection, operator, variables, ...)
}
\description{
Represents the intersection or union of variables within one collection.
}
\section{Slots}{

\describe{
\item{\code{collection}}{a categorical collection}

\item{\code{operator}}{"AND", "OR"}

\item{\code{variables}}{list of variables to be combined}
}}

\examples{
\dontrun{
cb <- CategoricalBatch("Title", "OR", c("The Matrix", "Casablanca"))
}

}
\seealso{
Useful links:
\itemize{
\item \url{https://code.tag.bio/docs/variable-object-reference}
The Categorical-Batch data-function from Tag.bio syntax documentation.
}
}
