% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_fundamentals.R
\name{CategoricalCollection}
\alias{CategoricalCollection}
\title{An S3 class representing the Tag.bio CategoricalCollection}
\usage{
CategoricalCollection(
  collection,
  collection_size,
  collection_entity_count,
  ...
)
}
\description{
An S3 class representing the Tag.bio CategoricalCollection
}
