% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagData.R
\name{get_parameters}
\alias{get_parameters}
\title{Get parameters from a tagData object}
\usage{
get_parameters(tag_data)
}
\arguments{
\item{tag_data}{tagData object}
}
\value{
list of parameters
}
\description{
This getter method returns the pass through parameters of a tagData object.
}
