% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagConnect.R
\name{summary.tagConnect}
\alias{summary.tagConnect}
\title{Create a table from a Tag.bio data product}
\usage{
\method{summary}{tagConnect}(object, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Connects to a data product via a tagConnect connection.
}
\examples{
# connect to local host, no API key required
tag_con <- tagConnect()

# pulls a data table from the data product
fc <- tbl(tag_con) |> collect()

}
