% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_parse.R
\name{tag_categorical_batch_op}
\alias{tag_categorical_batch_op}
\title{importFrom utils methods
importFrom methods is}
\usage{
tag_categorical_batch_op(op, batch_op)
}
\description{
importFrom utils methods
importFrom methods is
}
