% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagRmd.R
\name{tag_report}
\alias{tag_report}
\title{Tag.bio templates from R Markdown}
\usage{
tag_report(
  theme = "default-color",
  highlight = NULL,
  css = NULL,
  math = c("mathjax", "katex"),
  fig_retina = NULL,
  keep_md = FALSE,
  readme = FALSE,
  compressed_css = FALSE,
  logo = NULL,
  ...
)
}
\arguments{
\item{theme}{Character string to specify the document theme.
Currently supported themes are \code{"cayman"}, \code{"tactile"},
\code{"architect"}, \code{"leonids"}, and \code{"hpstr"}.}

\item{highlight}{Character string to specify the syntax highlight theme.
Supported values are \code{NULL} (use the Pandoc default), \code{kate},
\code{monochrome}, \code{breezeDark}, \code{espresso}, \code{zenburn},
\code{haddock}, and \code{tango}.}

\item{css}{Additional CSS file to be \strong{merged} in the document. It will
\strong{NOT} overwrite the CSS from the themes.}

\item{math}{The engine to render math expressions. Possible values are
\code{"mathjax"} and \code{"katex"}. The \code{katex} engines
supports offline (no internet connection) rendering of math expressions.}

\item{fig_retina}{The same argument in \code{\link[rmarkdown]{html_document}}
but with a different default value (\code{NULL} to disable
retina scaling)}

\item{keep_md}{Keep the markdown file generated by knitting.
This is the same argument in \code{\link[rmarkdown]{html_vignette}}.}

\item{readme}{Use this vignette as the package \file{README.md} file.
This is the same argument in \code{\link[rmarkdown]{html_vignette}}.}

\item{compressed_css}{Whether the compressed CSS should be used.}

\item{\dots}{Additional arguments passed to
\code{\link[rmarkdown]{html_document}}.}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}.
}
\description{
This function is a slight reworking of the
\code{\link[prettydoc]{html_pretty}} method used to create HTML documents
from R Markdown that work well within the Tag.bio platform.
}
\examples{
\dontrun{
doc <- system.file("rmarkdown", "templates", "tagbio",
  "skeleton", "skeleton.Rmd",
  package = "prettydoc"
)
rmarkdown::render(doc, html_tag(), output_dir = tempdir())
}

}
\author{
J Ireland, Yixuan Qiu (prettydoc)
}
