% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tagbio.R
\docType{class}
\name{tagbio.data-class}
\alias{tagbio.data-class}
\alias{tagbio.data}
\title{An S4 class representing data from the tag.bio platform}
\description{
The tagbio.data class holds tagbio.data as a data.frame attribute of the
object.  tagbio.data objects are typically created through a
\code{getTagData} query or created directly from the flux capacitor and
passed to a user defined method.  The tagbio.data object may optionally
have a list of parameters defined.  This list is typically populated by
the flux capacitor and is used to direct user defined methods.
}
\section{Slots}{

\describe{
\item{\code{data.frame}}{a data frame typically retrieved from tag.bio}

\item{\code{parameters}}{a list of parameters from tag.bio}
}}

