% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagbio.R
\docType{package}
\name{tagbio}
\alias{tagbio-package}
\alias{tagbio}
\title{tagbio: A package for working with the tag.bio analysis platform}
\description{
The tagbio package provides an interface for connecting to a tag.bio
server, discovering data products and accessing data.  The package allows
you to work with remote data products as if they are in-memory data frames.
The package follows the tidyverse design and should feel right at home to anyone
familiar with tidyverse libraries.
}
\section{tagbio classes}{

To be filled in ...
}

\seealso{
Useful links:
\itemize{
\item \url{https://tag.bio}
\item \url{https://code.tag.bio}
}
}
