% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tagbio.R
\docType{class}
\name{tagbio.result-class}
\alias{tagbio.result-class}
\alias{tagbio.result}
\title{An S4 class representing data to return to the tag.bio platform}
\description{
The tagbio.result class allows the user to return data to the tag.bio
platform.  It is typically only used in user-defined protocol methods
called from the flux capacitor.
}
\section{Slots}{

\describe{
\item{\code{data.frame}}{a dataframe of results}

\item{\code{jpeg}}{path to a jpeg file}

\item{\code{pdf}}{path to a pdf file}

\item{\code{png}}{path to a png file}

\item{\code{svg}}{path to a svg file}
}}

