% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagData.R
\name{tagData}
\alias{tagData}
\title{An S3 class representing data response from a protocol}
\usage{
tagData(results = data.frame(), parameters = list())
}
\arguments{
\item{results}{a data frame response from the protocol}

\item{parameters}{a list of pass-through parameters from the protocol}
}
\description{
The tagData class captures the resulting data frame and pass-through
arguments returned from a data product protocol.
}
\details{
The tagData class holds tagbio.data as a data.frame attribute of the
object.  TagbioData objects are typically created through a
\code{getTagData} query or created directly from the flux capacitor and
passed to a user defined method.  The TagbioData object may optionally
have a list of parameters defined.  This list is typically populated by
the flux capacitor and is used to direct user defined methods.
}
