% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagFC.R
\name{tagFC}
\alias{tagFC}
\title{An S3 class representing a data product (aka "FC") on a Tag.bio server}
\usage{
tagFC(con, fc = "", qdelim = " = ")
}
\description{
The tagFC class allows access to a Tag.bio data product as if it were an
in-memory object.  An instance of this class may be used to filter items,
select attributes and ultimately return a table of data from a Tag.bio data
product.
}
\section{Slots}{

\describe{
\item{\code{con}}{a tagConnect object}

\item{\code{fc}}{name of the data product.
When running locally, this should be left empty}

\item{\code{qdelim}}{delimiter character used to separate collection names
and collection values}

\item{\code{qselect}}{list of attribute selections.  Normally this should be set up
using \code{select} methods}

\item{\code{qfilter}}{list of item filters.  Normally this should be set up
using \code{filter} methods}

\item{\code{collection_defs}}{definitions of available collections for this data
product. Populated when a tagbio_fc object is first created.}

\item{\code{url}}{full URL to this data product}
}}

\section{Construction}{

\code{tagFC} objects are not normally created directly, but are instead
created using the \code{tbl} method, much like how a database connection
is made through \code{dbplyr}.
}

\section{Methods}{

\code{summary(tc)}: Shows all collections in this data product.  Results are
returned in a table with columns.
}

\section{Miscellaneous}{

Data products at tag.bio were historically called "flux capacitors"
or "FC"s for short.  The "fc" name lives on in some of the internal
classes and methods.
}

\examples{
\dontrun{
# connect to local host, no API key required
tag_con <- tagConnect()

# access to tagFC object from local instance
fc <- tbl(tag_con)
}

}
\seealso{
Useful links:
\itemize{
\item \url{https://backtothefuture.fandom.com/wiki/Flux_capacitor}
for the origin of "flux capacitor"
}
}
