% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagConnect.R
\name{tagListFCs.tagConnect}
\alias{tagListFCs.tagConnect}
\title{List all Tag.bio data products using a tagConnect connection}
\usage{
\method{tagListFCs}{tagConnect}(.data)
}
\description{
Connects to a data product via a tagConnect connection.
}
