% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagConnect.R
\name{tbl.tagConnect}
\alias{tbl.tagConnect}
\title{Create a table from a tag.bio data product}
\usage{
\method{tbl}{tagConnect}(src = "tagConnect", fc = "")
}
\arguments{
\item{src}{A data source}
}
\description{
Connects to a data product via a tagConnect connection.
}
\examples{
# connect to local host, no API key required
tag_con <- tagConnect()

# pulls a data table from the data product
fc <- tbl(tag_con) \%>\% collect()


}
