% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagResult.R
\name{add_result}
\alias{add_result}
\title{Add data to a tagResult object}
\usage{
add_result(tag_result, results, result_type)
}
\arguments{
\item{tag_result}{tagResult object}

\item{results}{data.frame or file path to results data}

\item{result_type}{either \code{data.frame}, \code{jpeg}, \code{pdf},
\code{png}, \code{svg}}

\item{tag_data}{tagData object}
}
\value{
updated tagResult object
}
\description{
This method adds results to a tagResult object
}
\seealso{
\code{\link{tagData}}
}
