% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tagbio.R
\name{addResult}
\alias{addResult}
\title{Add data to a tagbio.result.}
\usage{
addResult(tagResult, resultData, resultType)
}
\arguments{
\item{tagResult}{tagResult object}

\item{resultData}{data.frame or file path to results data}

\item{resultType}{either \code{data.frame}, \code{jpeg}, \code{pdf},
\code{png}, \code{svg}}
}
\value{
updated tagResult object
}
\description{
This method adds results to a tagbio.result object.
}
\examples{
addResult(tagresult, "my_pdf_file.pdf", type = "pdf")
}
\seealso{
\code{\link{tagbio.data}}
}
