% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tagbio.R
\name{addResult}
\alias{addResult}
\title{Add data to a TagbioResult.}
\usage{
addResult(tag_result, result_data, result_type)
}
\arguments{
\item{tag_result}{tagResult object}

\item{result_data}{data.frame or file path to results data}

\item{result_type}{either \code{data.frame}, \code{jpeg}, \code{pdf},
\code{png}, \code{svg}}
}
\value{
updated TagbioResult object
}
\description{
This method adds results to a TagbioResult object
}
\examples{
tag_result <- TagbioResult()
addResult(tag_result, "my_pdf_file.pdf", result_type = "pdf")
}
\seealso{
\code{\link{TagbioData}}
}
