% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_fundamentals.R
\name{CategoricalBatch}
\alias{CategoricalBatch}
\title{An S3 class representing a CategoricalBatch}
\usage{
CategoricalBatch(collection, operator, variables, ...)
}
\description{
Represents an AND or OR of variables in the collection
}
\section{Slots}{

\describe{
\item{\code{collection}}{a categorical collection}

\item{\code{operator}}{"AND", "OR"}

\item{\code{variables}}{list of variables to be combined}
}}

\examples{
cb <- CategoricalBatch("Title", "OR", c("The Matrix", "Casablanca"))

}
\seealso{
Useful links:
\itemize{
\item \url{https://code.tag.bio/docs/variable-object-reference} tag.bio documentation
}
}
