% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_fundamentals.R
\name{CategoricalCompound}
\alias{CategoricalCompound}
\title{An S3 class representing the Tag.bio CategoricalCompound}
\usage{
CategoricalCompound(criteria = list(), operator, ...)
}
\description{
A categorical-compound variable object converts two or more
categorical variable objects representing variables into the intersection
or the union of those two sets.
}
\details{
If the operator attribute is AND, the categorical variable objects
in the criteria attribute will be intersected.
}
\section{Slots}{

\describe{
\item{\code{criteria}}{list of categorical variable objects}

\item{\code{operator}}{"AND", "OR"}
}}

\examples{
cc <- CategoricalCompound(c(), "AND")

}
\seealso{
Useful links:
\itemize{
\item \url{https://code.tag.bio/docs/variable-object-reference#categoricalcompound}
The Categorical-Compound data-function from Tag.bio syntax documentation.
}
}
