% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_fundamentals.R
\name{CategoricalVariable}
\alias{CategoricalVariable}
\title{An S3 class representing a CategoricalVariable}
\usage{
CategoricalVariable(collection, variable, ...)
}
\description{
An S3 class representing a CategoricalVariable
}
\section{Slots}{

\describe{
\item{\code{collection}}{a categorical collection}

\item{\code{variable}}{a categorical variable}
}}

\examples{
cvr <- CategoricalVariable("color", "red")

}
\seealso{
Useful links:
\itemize{
\item \url{https://code.tag.bio/docs/variable-object-reference#categorical} tag.bio documentation
}
}
