% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagData.R
\name{get_results}
\alias{get_results}
\title{Get result data from a tagData object}
\usage{
get_results(tag_data, data_type = NA, row_name = "")
}
\arguments{
\item{tag_data}{tagData object}

\item{data_type}{type of data to return in data graphics::frame (default NA)}

\item{row_name}{if set, use this column to create row names (default "")}
}
\value{
data.frame of data
}
\description{
This getter method returns the data.frame component of a tagData object.
If a data_type is provided, only results of this type are returned in
the data.frame.
}
