% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tagbio.R
\name{getTagData}
\alias{getTagData}
\title{Retrieve data from a tag.bio flux capacitor}
\usage{
getTagData(fcUrl, apiKey, arguments, protocolName = "download")
}
\arguments{
\item{fcUrl}{flux capacitor API URL}

\item{apiKey}{flux capacitor API key for authentication}

\item{arguments}{a list of arguments to specify the API query}

\item{protocolName}{name of protocol.  Default is "download".}
}
\value{
A tagbio.data object with data.frame populated by query
}
\description{
Retrieve data from a tag.bio flux capacitor
}
\examples{
getTagData(FC_URL, API_KEY, list(expression = c("A1CF","A2M"),
    clinical_categorical_variables = c("clinical.SAMPLE_ID")))

}
\seealso{
\code{\link{tagbio.result}},\code{\link{getTagData}}
}
