% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tagbio.R
\name{getTagData}
\alias{getTagData}
\title{Retrieve data from a tag.bio flux capacitor}
\usage{
getTagData(
  fc,
  protocol_instance = NULL,
  script = NULL,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{fc}{flux capacitor object}

\item{protocol_instance}{protocol instance}

\item{script}{script instance}
}
\value{
A tagbio.data object with data.frame populated by query
}
\description{
Retrieve data from a tag.bio flux capacitor
}
\seealso{
\code{\link{TagbioResult}},\code{\link{FC}},\code{\link{ProtocolInstance}}
}
