% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_fundamentals.R
\name{NumericSlice}
\alias{NumericSlice}
\title{An S3 class representing a NumericSlice}
\usage{
NumericSlice(
  criterion,
  operator,
  value = numeric(),
  percentile = numeric(),
  ...
)
}
\description{
An S3 class representing a NumericSlice
}
\section{Slots}{

\describe{
\item{\code{criterion}}{a numeric variable being compared}

\item{\code{operator}}{"<", "<=", "==", "!=", ">=", ">"}

\item{\code{value}}{numeric comparator}

\item{\code{percentile}}{percentile comparator}
}}

\examples{
ns <- NumericSlice("carat", ">", 0.3)

}
\seealso{
Useful links:
\itemize{
\item \url{https://code.tag.bio/docs/variable-object-reference#numericslice} tag.bio documentation
}
}
