% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_fundamentals.R
\name{NumericVariable}
\alias{NumericVariable}
\title{An S3 class representing tag.bio NumericVariable}
\usage{
NumericVariable(collection, variable, ...)
}
\description{
An S3 class representing tag.bio NumericVariable
}
