% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagConnect.R
\name{tagListFCs}
\alias{tagListFCs}
\title{An S3 class to list all data products on a Tag.bio server}
\usage{
tagListFCs(.data)
}
\arguments{
\item{.data}{tagConnect object}
}
\description{
The tagConnect class establishes a connection to a local or remote
Tag.bio server.
}
