% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagResult.R
\name{tagResult}
\alias{tagResult}
\title{An S3 class representing data to return to the tag.bio platform}
\usage{
tagResult(
  output_path,
  result_type,
  results = tibble::tibble(),
  message_path = NA
)
}
\arguments{
\item{output_path}{path to results file}

\item{result_type}{format of the results (html, png, pdf)}

\item{message_path}{path to a message file}

\item{result_data}{a data frame of results}
}
\description{
The tagResult class allows the user to return data to the tag.bio
platform.  It is typically only used in user-defined protocol methods
called from the flux capacitor.
}
\details{
Results are typically stored in a file either as a data frame or a
plot saved in a graphics format or HTML.  Any messages not part of
the results can be placed in the message file.

A tagResult is typically pre-populated and passed to a user-
defined R function.  The type and paths direct the user what output
is expected and where results should be saved.
}
