% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagConnect.R
\name{tbl.tagConnect}
\alias{tbl.tagConnect}
\title{Create a table from a Tag.bio data product}
\usage{
\method{tbl}{tagConnect}(src = "tagConnect", fc = "")
}
\arguments{
\item{src}{tagConnect object}

\item{fc}{data product name}
}
\description{
Connects to a data product via a tagConnect connection.
}
\examples{
\dontrun{
#connect to local host, no API key required
tag_con <- tagConnect()

#pulls a data table from the data product
fc <- tbl(tag_con) |> collect()
}

}
