% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_fundamentals.R
\name{to_json}
\alias{to_json}
\title{Translate tag data constructs as JSON}
\usage{
to_json(.Object, ...)
}
\description{
The \code{to_json} method converts tag data constructs to JSON, typically for
communicating with the API or creating configuration files.
}
\section{Slots}{

\describe{
\item{\code{.Object}}{a tag data object}
}}

\examples{
# TODO

}
